Test cases for path expressions need to inherit from requires_server_version automatically.
